requirements/test.txt
aioresponses==0.1.4
freezegun==0.3.8
pytest==3.0.4
pytest-aiohttp==0.1.3
pytest-flakes==1.0.1
pytest-mock==1.6.0
pytest-pep8==1.0.6
pytest-capturelog==0.7
mypy==0.521
# TODO: unlock after publish this packages
pytest-async-fixtures==0.1.6
pytest-peewee-async==0.2.0
